[package][mediacentar-addon-osmc] Networking: added automatic check for Wait_for_Network
New process, when the Networking module first starts, it scans the
advancedsettings.xml file and determines if it contains valid MySQL
information. If it does, it forces the Wait_for_Network setting to be
true. This is required as there will be people who simply restore an old
copy of their advancedsettings.xml without opening the GUI.